ceca/requirements.txt
django-browser-reload==1.10.0
  Django==4.2.2
    asgiref==3.7.2
      typing_extensions==4.6.3
    sqlparse==0.4.4
django-on-heroku==1.1.2
  dj-database-url==2.0.0
    Django==4.2.2
      asgiref==3.7.2
        typing_extensions==4.6.3
      sqlparse==0.4.4
    typing_extensions==4.6.3
  Django==4.2.2
    asgiref==3.7.2
      typing_extensions==4.6.3
    sqlparse==0.4.4
  psycopg2-binary==2.9.6
  whitenoise==6.5.0
django-storages==1.13.2
  Django==4.2.2
    asgiref==3.7.2
      typing_extensions==4.6.3
    sqlparse==0.4.4
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
  Django==4.2.2
    asgiref==3.7.2
      typing_extensions==4.6.3
    sqlparse==0.4.4
  pytz==2023.3
factory-boy==3.2.1
  Faker==18.11.2
    python-dateutil==2.8.2
      six==1.16.0
gunicorn==20.1.0
  setuptools==68.0.0
jmespath==1.0.1
loguru==0.7.0
Pillow==9.5.0
pip==23.1.2
pipdeptree==2.9.3
psycopg2==2.9.6
python-dotenv==1.0.0
urllib3==2.0.3
watchfiles==0.19.0
  anyio==3.7.0
    exceptiongroup==1.1.1
    idna==3.4
    sniffio==1.3.0
wheel==0.40.0
